More client refactoring
