+++ servidor.c
@@ -12,26 +12,31 @@
 #include <unistd.h>
 
 #include "discovery.h"
+#include "utils.h"
 
 #define PORT 4000
 #define MAX_CLIENTS 100
 #define PACKET_TYPE_REQ 2
 #define PACKET_TYPE_REQ_ACK 4
 
-struct requisicao {
+struct requisicao
+{
     uint32_t value;
 };
 
-struct requisicao_ack {
+struct requisicao_ack
+{
     uint32_t seqn;
     uint32_t num_reqs;
     uint64_t total_sum;
 };
 
-typedef struct __packet {
+typedef struct __packet
+{
     uint16_t type;
     uint32_t seqn;
-    union {
+    union
+    {
         struct requisicao req;
         struct requisicao_ack ack;
     } data;
@@ -50,8 +55,10 @@ uint32_t total_reqs = 0;
 uint64_t total_sum = 0;
 pthread_mutex_t lock = PTHREAD_MUTEX_INITIALIZER;
 
-int find_or_add_client(struct sockaddr_in *addr) {
-    for (int i = 0; i < client_count; i++) {
+int find_or_add_client(struct sockaddr_in *addr)
+{
+    for (int i = 0; i < client_count; i++)
+    {
         if (clients[i].addr.sin_addr.s_addr == addr->sin_addr.s_addr)
             return i;
     }
@@ -61,12 +68,8 @@ int find_or_add_client(struct sockaddr_in *addr) {
     return client_count++;
 }
 
-void current_time(char *buf, size_t len) {
-    time_t now = time(NULL);
-    strftime(buf, len, "%Y-%m-%d %H:%M:%S", localtime(&now));
-}
-
-void *handle_request(void *arg) {
+void *handle_request(void *arg)
+{
     struct
     {
         packet pkt;
@@ -80,7 +83,8 @@ void *handle_request(void *arg) {
     int idx = find_or_add_client(&ctx->addr);
 
     pthread_mutex_lock(&lock);
-    if (pkt->seqn == clients[idx].last_req + 1) {
+    if (pkt->seqn == clients[idx].last_req + 1)
+    {
         total_reqs++;
         total_sum += pkt->data.req.value;
         clients[idx].last_req = pkt->seqn;
@@ -89,7 +93,9 @@ void *handle_request(void *arg) {
         printf("%s client %s id_req %u value %u num_reqs %u total_sum %lu\n",
                timebuf, inet_ntoa(ctx->addr.sin_addr), pkt->seqn,
                pkt->data.req.value, total_reqs, total_sum);
-    } else {
+    }
+    else
+    {
         current_time(timebuf, sizeof(timebuf));
         printf("%s client %s DUP!! id_req %u value %u num_reqs %u total_sum %lu\n",
                timebuf, inet_ntoa(ctx->addr.sin_addr), pkt->seqn,
@@ -110,8 +116,10 @@ void *handle_request(void *arg) {
     return NULL;
 }
 
-int main(int argc, char *argv[]) {
-    if (argc != 2) {
+int main(int argc, char *argv[])
+{
+    if (argc != 2)
+    {
         fprintf(stderr, "Uso: %s <porta>\n", argv[0]);
         exit(EXIT_FAILURE);
     }
@@ -121,7 +129,8 @@ int main(int argc, char *argv[]) {
 
     // Cria o socket com domínio AF_INET e tipo SOCK_DGRAM (protocolo UDP)
     int sock;
-    if ((sock = socket(AF_INET, SOCK_DGRAM, 0)) == -1) {
+    if ((sock = socket(AF_INET, SOCK_DGRAM, 0)) == -1)
+    {
         perror("Cannot ceate socket");
         exit(EXIT_FAILURE);
     }
@@ -129,13 +138,14 @@ int main(int argc, char *argv[]) {
     // Cria o container genérico que permite que o SO identifique a família do endereço
     struct sockaddr_in servaddr;
     memset((char *)&servaddr, 0, sizeof(servaddr));
-    servaddr.sin_family = AF_INET;                 // IPv4
-    servaddr.sin_port = htons(port);               // Porta do servidor
-    servaddr.sin_addr.s_addr = htonl(INADDR_ANY);  // Aceita qualquer endereço
+    servaddr.sin_family = AF_INET;                // IPv4
+    servaddr.sin_port = htons(port);              // Porta do servidor
+    servaddr.sin_addr.s_addr = htonl(INADDR_ANY); // Aceita qualquer endereço
     bzero(&(servaddr.sin_zero), 8);
 
     // Vincular o socket à porta
-    if (bind(sock, (struct sockaddr *)&servaddr, sizeof(servaddr)) < 0) {
+    if (bind(sock, (struct sockaddr *)&servaddr, sizeof(servaddr)) < 0)
+    {
         perror("Cannot bind socket");
         close(sock);
         exit(EXIT_FAILURE);
@@ -145,16 +155,20 @@ int main(int argc, char *argv[]) {
     current_time(timebuf, sizeof(timebuf));
     printf("%s num_reqs 0 total_sum 0\n", timebuf);
 
-    while (1) {
+    while (1)
+    {
         packet pkt;
         struct sockaddr_in cliaddr;
         socklen_t len = sizeof(cliaddr);
 
         recvfrom(sock, &pkt, sizeof(pkt), 0, (struct sockaddr *)&cliaddr, &len);
 
-        if (pkt.type == PACKET_TYPE_DISCOVERY) {
+        if (pkt.type == PACKET_TYPE_DISCOVERY)
+        {
             discovery_handle_request(sock, &cliaddr, len, (discovery_packet *)&pkt);
-        } else if (pkt.type == PACKET_TYPE_REQ) {
+        }
+        else if (pkt.type == PACKET_TYPE_REQ)
+        {
             pthread_t tid;
             void *ctx = malloc(sizeof(packet) + sizeof(cliaddr) + sizeof(socklen_t) + sizeof(int));
             memcpy(ctx, &(struct { packet pkt; struct sockaddr_in addr; socklen_t addrlen; int sock; }){pkt, cliaddr, len, sock}, sizeof(packet) + sizeof(cliaddr) + sizeof(socklen_t) + sizeof(int));
