+++ utils.h
@@ -1,0 +1,3 @@
+#include <stdio.h>
+
+void current_time(char *buf, size_t len);
